+++ system-tests/debug-logging-default-test/src/test/java/com/teamscale/client/DebugLoggingDefaultTest.java
@@ -2,19 +2,32 @@ package com.teamscale.client;
 
 import org.junit.jupiter.api.Test;
 
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.Optional;
+import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 public class DebugLoggingDefaultTest {
 
-	private static final Path LOG_DIRECTORY = Paths.get("../../agent/build", "logs");
-
 	@Test
 	public void systemTest() throws Exception {
-		assertThat(Files.exists(LOG_DIRECTORY)).isTrue();
-		assertThat(Files.readAllLines(LOG_DIRECTORY.resolve("teamscale-jacoco-agent.log"))).isNotEmpty();
+		Path logDirectory = searchForAgentTempDirectory().resolve("logs");
+		assertThat(logDirectory).exists();
+		assertThat(Files.readAllLines(logDirectory.resolve("teamscale-jacoco-agent.log"))).isNotEmpty();
 	}
+
+	private Path searchForAgentTempDirectory() throws IOException {
+		Path tempDirectory = Paths.get(System.getProperty("java.io.tmpdir"));
+		try (Stream<Path> stream = Files.list(tempDirectory)) {
+			Optional<Path> maybeTempDirectory = stream.
+					filter(path -> path.getFileName().toString().contains("teamscale-java-profiler"))
+					.sorted().findFirst();
+			return maybeTempDirectory.orElseThrow(() -> new AssertionError("Could not locate agent temp directory"));
+		}
+	}
+
 }